requirements.txt
fastapi[standard]
sqlalchemy==2.0.36
psycopg2==2.9.10
pyjwt==2.10.0
passlib==1.7.4
httpx==0.27.2
pytest==8.3.3
python-dotenv==1.0.1
# mangum
# boto3==1.35.55